requirements.txt
discord.py
asyncio
Pillow
python-dotenv
numpy
